refactor(crates): separate wrapper crate & CLI
