+++ Cargo.toml
@@ -4,13 +4,14 @@ version = "0.2.0"
 license = "MIT OR Apache-2.0"
 description = "Havok Behavior (De)Serialization Library(ver. hk_2010.2.0-r1)"
 repository = "https://github.com/SARDONYX-sard/serde-hkx"                     # NOTE: Without it, CLI & CI build will error.
-rust-version = "1.79"                                                         # For count_bytes method
+rust-version = "1.79"
 edition = "2021"
 
 [workspace]
 members = [
   "crates/class_generator",
   "crates/cli",
+  "crates/serde_hkx_features",
   "crates/havok_classes",
   "crates/havok_serde",
   "crates/havok_types",
@@ -23,35 +24,38 @@ default-members = ["crates/cli"]
 resolver = "2"
 
 [workspace.dependencies]
-bitflags = "2.4"                                               # To generate flags type
-educe = "0.6.0"
-indexmap = "2.2.6"
+bitflags = "2.6.0"                                             # To generate flags type
+color-print = "0.3.6"                                          # To color example help
+educe = "0.6.0"                                                # Custom derive by derive
+indexmap = "2.5.0"                                             # Ordering HashMap
 jwalk = { version = "0.8" }                                    # To enumarate json class files
 num-derive = "0.4"                                             # To serialize Enum & Flags
-num-traits = "0.2"                                             # To serialize Enum & Flags
+num-traits = "0.2.19"                                          # To serialize Enum & Flags
 parse-display = "0.10.0"                                       # Display derive
 parse_int = { version = "0.6" }
-proc-macro2 = { version = "1.0.82", default-features = false }
-quote = { version = "1.0.36", default-features = false }
+proc-macro2 = { version = "1.0.82", default-features = false } # Rust token
+quote = { version = "1.0.37", default-features = false }       # Rust token gen to more convinience
 serde = { version = "1.0", features = ["derive"] }             # Implement (De)Serializer
-serde_json = { version = "1.0.116" }                           # To read and write Json files
-serde_with = "3.8.1"                                           # Serde utility
+serde_json = { version = "1.0.128" }                           # To read and write Json files
+serde_with = "3.9.0"                                           # Serde utility
 similar = "2.6.0"                                              # Create diff
-snafu = "0.8.2"                                                # Define all error patterns(with backtrace)
-static_assertions = "1.1.0"
-syn = { version = "2.0.63", default-features = false }
-tokio = { version = "1.37.0" }
-tracing = { version = "0.1.40" }
+snafu = "0.8.4"                                                # Define all error patterns(with backtrace)
+static_assertions = "1.1.0"                                    # Compile tim assertions
+syn = { version = "2.0.77", default-features = false }         # Rust parser
+tokio = { version = "1.40.0" }                                 # Async runtime
+tracing = { version = "0.1.40" }                               # logger
 
 # workspace members
-havok_classes = { path = "./crates/havok_classes", default-features = false }
+havok_classes = { path = "./crates/havok_classes" }
 havok_serde = { path = "./crates/havok_serde" }
 havok_types = { path = "./crates/havok_types" }
 havok_types_derive = { path = "./crates/havok_types_derive" }
+serde_hkx = { path = "./serde_hkx", default-features = false }
+serde_hkx_features = { path = "./crates/serde_hkx_features" }
 
 # dev-dependencies
 pretty_assertions = "1.4"                                    # Color diff assertion
-quick_tracing = { version = "0.1.5", features = ["derive"] }
+quick_tracing = { version = "0.1.5", features = ["derive"] } # Easy logging
 
 [profile.release]
 lto = "thin"
