+++ README.md
@@ -4,30 +4,41 @@
 
 [CLI release page](https://github.com/SARDONYX-sard/serde-hkx/releases)
 
-The following functions can be attempted using the CLI.
-
-- 32bit to 64bit (reverse) conversion of hkx.
-- XML to 32bit/64bit hkx (reverse) conversion.
-- Display of hkx/XML state machine dependency tree.
-- Output logs that make it possible to determine the binary data location of hkx.
-- hexdump the binary data of hkx.
-- Display data differences between two hkx/XML.
+- Features
+  - [x] 32bit to 64bit (reverse) conversion of hkx.
+  - [x] XML to 32bit/64bit hkx (reverse) conversion.
+  - [x] Display of hkx/XML state machine dependency tree.
+  - [x] Output logs that make it possible to determine the binary data location of hkx.
+  - [x] Hexdump the binary data of hkx.
+  - [x] Display data differences between two hkx/XML.
 
 ## Usage(As a library)
 
-- Cargo.toml
+- Convenience wrapper API(For CLI/GUI): (e.g. [CLI](./crates/cli/src/cli/mod.rs))
 
 ```toml
-havok_classes = { git = "https://github.com/SARDONYX-sard/serde-hkx" }
-havok_serde = { git = "https://github.com/SARDONYX-sard/serde-hkx" }
-havok_types = { git = "https://github.com/SARDONYX-sard/serde-hkx" }
-serde_hkx = { git = "https://github.com/SARDONYX-sard/serde-hkx" }
+# in Cargo.toml
+serde_hkx_features = { git = "https://github.com/SARDONYX-sard/serde-hkx", tag = "0.2.0" }
 ```
 
-Examples
+- Low level API: (e.g. [Tests](./serde_hkx/src/tests/verify.rs))
 
-- [Tests](./serde_hkx/src/tests/verify.rs)
-- [Convert(in cli)](./crates/cli/src/cli/convert.rs)
+```toml
+# in Cargo.toml
+havok_classes = { git = "https://github.com/SARDONYX-sard/serde-hkx", tag = "0.2.0" }
+serde_hkx = { git = "https://github.com/SARDONYX-sard/serde-hkx", tag = "0.2.0" }
+```
+
+## Documentation
+
+- Specification: [specification](./docs/specification/hkx_binary_format.md)
+- Binary analysis: [hkx spec](./docs/handson_hex_dump/defaultmale/readme.md)
+
+- API:
+
+```shell
+git clone https://github.com/SARDONYX-sard/serde-hkx.git;cd ./serde-hkx;cargo docs --open;
+```
 
 ## Try implementations
 
@@ -44,11 +55,10 @@ The only way to find out if it is possible or not is to try it out.
 - [x] Generate Classes
 
 - [x] Testing the API & CI.
-- [ ] Modify the code based on the test data. <- (Finally, only global_fixup order issues...)
+- [x] Modify the code based on the test data. (since v0.2.0)
 
 ## About the GPL old version
 
 GPL-dependent code (`rhexdump` crate) has been removed, and the GPL usage history has been completely removed.
-Previous GPL-tainted versions can be found at the following links.
 
 - [Prev GPL History(ver. 0.0.0)](https://github.com/SARDONYX-sard/serde-hkx/releases/tag/0.0.0-prev-gpl-history)
