+++ crates/cli/src/args/color.rs
