+++ crates/cli/src/args/dump.rs
@@ -1,11 +1,5 @@
 //! Dump binary data in hexadecimal
-use crate::{
-    error::{Error, Result},
-    read_ext::ReadExt as _,
-};
-use serde_hkx::bytes::hexdump;
-use std::path::{Path, PathBuf};
-use tokio::fs;
+use std::path::PathBuf;
 
 /// ANSI color representation command examples.
 pub const EXAMPLES: &str = color_print::cstr!(
@@ -35,34 +29,4 @@ NOTE: Bytes cannot be written to stdout, so always use `-o` to specify the outpu
     /// Reverse conversion from hexdump to bytes
     #[clap(short, long)]
     pub reserve: bool,
-}
-
-/// Output bytes to stdout/file.
-pub async fn to_bytes<I, O>(input: I, output: Option<O>) -> Result<()>
-where
-    I: AsRef<Path>,
-    O: AsRef<Path>,
-{
-    match output {
-        Some(output) => {
-            let bytes = hexdump::to_bytes(&input.read_any_string().await?);
-            fs::write(output, &bytes).await?
-        }
-        None => return Err(Error::InvalidStdout),
-    };
-    Ok(())
-}
-
-/// Output hexdump to stdout/file.
-pub async fn to_string<I, O>(input: I, output: Option<O>) -> Result<()>
-where
-    I: AsRef<Path>,
-    O: AsRef<Path>,
-{
-    let hexdump = hexdump::to_string(input.read_bytes().await?); // NOTE: With newline.
-    match output {
-        Some(output) => fs::write(output, &hexdump).await?,
-        None => print!("{hexdump}"),
-    };
-    Ok(())
 }
