+++ crates/readme.md
@@ -1,0 +1,17 @@
+# crate dependency tree
+
+More details: Use `cargo tree` command.
+
+```txt
+cli/gui
+└── serde_hkx_features
+    ├── havok_classes <- (generated by class_generator)
+    │   ├── havok_serde
+    │   │   └── havok_types
+    │   ├── havok_types
+    │   └── havok_types_derive
+    │       └── havok_types_derive_proc_macros
+    └─ serde_hkx
+        ├── havok_serde
+        └── havok_types
+```
