+++ crates/serde_hkx_features/src/diff.rs
@@ -1,41 +1,11 @@
 use crate::error::{Error, Result};
 use crate::read_ext::ReadExt as _;
 use serde_hkx::bytes::hexdump;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use tokio::fs;
 
-/// ANSI color representation command examples.
-pub const EXAMPLES: &str = color_print::cstr!(
-    r#"<blue><bold><underline>Examples</underline></bold></blue>
-- <blue!>XML diff -> stdout</blue!>
-  <cyan!>hkxc diff</cyan!> ./defaultmale_by_hkxcmd.xml ./defaultmale_by_HavokBehaviorProcess.xml
-
-- <blue!>hkx hexdump diff -> stdout</blue!>
-  <cyan!>hkxc diff</cyan!> ./defaultmale_by_hkxcmd.hkx ./defaultmale_by_HavokBehaviorProcess.hkx
-
-- <blue!>hexdump diff -> a file</blue!>
-  <cyan!>hkxc diff</cyan!> ./defaultmale_x86_hexdump.txt ./defaultmale_x64_hexdump.txt <cyan!>-o</cyan!> diff.txt
-"#
-);
-
-#[derive(Debug, clap::Args)]
-#[clap(arg_required_else_help = true, after_long_help = EXAMPLES)]
-pub(crate) struct Args {
-    /// Old path
-    pub old: PathBuf,
-    /// New path
-    pub new: PathBuf,
-
-    /// Output with ANSI color (For standard output)
-    #[clap(long)]
-    pub color: bool,
-
-    /// If specified, write to a file (If not specified, stdout)
-    #[clap(short, long)]
-    pub output: Option<PathBuf>,
-}
-
 /// Show diff between two files.
+/// - `use_color`: ANSI color diff. (red & green)
 pub async fn exec<I1, I2, O>(old: I1, new: I2, output: Option<O>, use_color: bool) -> Result<()>
 where
     I1: AsRef<Path>,
@@ -76,7 +46,10 @@ where
     }
 }
 
-fn diff(old: impl AsRef<str>, new: impl AsRef<str>, color: bool) -> String {
+/// Show diff between two files.
+///
+/// - `color`: ANSI color diff. (red & green)
+pub fn diff(old: impl AsRef<str>, new: impl AsRef<str>, color: bool) -> String {
     let mut output_diff = String::new();
 
     let diff = similar::TextDiff::from_lines(old.as_ref(), new.as_ref());
