+++ crates/serde_hkx_features/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod convert;
+pub mod diff;
+pub mod dump;
+pub mod error;
+pub mod read_ext;
+pub mod tree;
+
+pub(crate) type ClassMap<'a> = indexmap::IndexMap<usize, havok_classes::Classes<'a>>;
