+++ crates/serde_hkx_features/src/tree.rs
@@ -6,36 +6,10 @@ use crate::{
 };
 use serde_hkx::{from_bytes, from_str, tree::HavokTree as _};
 use snafu::ResultExt as _;
-use std::{
-    io::Read as _,
-    path::{Path, PathBuf},
-};
+use std::{io::Read as _, path::Path};
 
-/// ANSI color representation command examples.
-pub const EXAMPLES: &str = color_print::cstr!(
-    r#"<blue><bold><underline>Examples</underline></bold></blue>
-- <blue!>xml -> dependencies tree -> stdout</blue!>
-  <cyan!>hkxc tree</cyan!> ./defaultmale.xml
-
-- <blue!>hkx -> dependencies tree -> stdout(+log)</blue!>
-  <cyan!>hkxc tree</cyan!> ./defaultmale.hkx <cyan!>--log-level</cyan!> trace <cyan!>--log-file</cyan!> ./hkx_tree.log
-
-- <blue!>hkx -> dependencies tree -> a file</blue!>
-  <cyan!>hkxc tree</cyan!> ./defaultmale.hkx <cyan!>-o</cyan!> tree.txt
-"#
-);
-
-#[derive(Debug, clap::Args)]
-#[clap(arg_required_else_help = true, after_long_help = EXAMPLES)]
-pub(crate) struct Args {
-    /// hkx/xml file path
-    pub input: PathBuf,
-    /// If specified, write to a file (If not specified, stdout)
-    #[clap(short, long)]
-    pub output: Option<PathBuf>,
-}
-
-/// Output tree to stdout/file.
+/// Output reference tree to stdout/file.
+/// - `output`: If not provided, then stdout.
 pub async fn output<I, O>(input: I, output: Option<O>) -> Result<()>
 where
     I: AsRef<Path>,
@@ -49,8 +23,8 @@ where
     Ok(())
 }
 
-/// Generate tree.
-async fn gen<P>(input: P) -> Result<String>
+/// Generate reference tree.
+pub async fn gen<P>(input: P) -> Result<String>
 where
     P: AsRef<Path>,
 {
