+++ serde_hkx/src/bytes/de/mod.rs
@@ -893,10 +893,10 @@ mod tests {
     }
 
     #[test]
-    #[cfg_attr(
-        feature = "tracing",
-        quick_tracing::init(test = "deserialize_hkx_bytes", stdio = false)
-    )]
+    // #[cfg_attr(
+    //     feature = "tracing",
+    //     quick_tracing::init(test = "deserialize_hkx_bytes", stdio = false)
+    // )]
     fn test_deserialize_class_index() {
         use havok_classes::Classes;
